Add files via upload
for the Advanced Excel functions, a lot of work was done on different excel functions. i made use of the following functions on excel to carry out the work; formula fun damentals, tax rate, order of operations, basic functions, conditional functions, text functions, logical functions, lookup functions, date and time functions and finding of formulas on the tables.
on the lookup function, vlookup and xlookup was used, in the logical functions; i applied IF, AND, OR and nested functions to perform logical operations and automate decision making. while i was working with dates and time, i learn how to manupulate format date/ time values for better data analysis.